+++ core/templates/dev/head/components/common-layout-directives/common-elements/attribution-guide.directive.ts
@@ -1,4 +1,4 @@
-// Copyright 2015 The Oppia Authors. All Rights Reserved.
+// Copyright 2016 The Oppia Authors. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,29 +13,28 @@
 // limitations under the License.
 
 /**
- * @fileoverview Directive for the background banner.
+ * @fileoverview Directive for the attribution guide.
  */
 
+require('domain/utilities/BrowserCheckerService.ts');
 require('domain/utilities/UrlInterpolationService.ts');
 
-oppia.directive('backgroundBanner', [
+oppia.directive('attributionGuide', [
   'UrlInterpolationService', function(UrlInterpolationService) {
     return {
       restrict: 'E',
       scope: {},
       bindToController: {},
       templateUrl: UrlInterpolationService.getDirectiveTemplateUrl(
-        '/components/background/background_banner_directive.html'),
+        '/components/common-layout-directives/common-elements/' +
+        'attribution-guide.directive.html'),
       controllerAs: '$ctrl',
       controller: [
-        function() {
+        'BrowserCheckerService', 'UrlService', function(
+            BrowserCheckerService, UrlService) {
           var ctrl = this;
-          var possibleBannerFilenames = [
-            'bannerA.svg', 'bannerB.svg', 'bannerC.svg', 'bannerD.svg'];
-          var bannerImageFilename = possibleBannerFilenames[
-            Math.floor(Math.random() * possibleBannerFilenames.length)];
-          ctrl.bannerImageFileUrl = UrlInterpolationService.getStaticImageUrl(
-            '/background/' + bannerImageFilename);
+          ctrl.isMobileDevice = BrowserCheckerService.isMobileDevice();
+          ctrl.iframed = UrlService.isIframed();
         }
       ]
     };
