+++ core/templates/dev/head/components/common-layout-directives/common-elements/confirmation-modal.template.html
