+++ core/templates/dev/head/components/common-layout-directives/common-elements/promo-bar.directive.html
