+++ core/templates/dev/head/domain/statistics/PlaythroughImprovementCardObjectFactory.ts
@@ -40,7 +40,7 @@ oppia.factory('PlaythroughImprovementCardObjectFactory', [
       var discardThis = function() {
         return $uibModal.open({
           templateUrl: UrlInterpolationService.getDirectiveTemplateUrl(
-            '/components/confirmation_modal_directive.html'),
+            '/components/common-layout-directives/common-elements/confirmation-modal.template.html'),
           backdrop: true,
           controller: [
             '$scope', '$uibModalInstance',
