+++ core/templates/dev/head/pages/contact/Contact.ts
@@ -16,5 +16,5 @@
  * @fileoverview File to import necessary scripts for contact page.
  */
 
-require('components/background/BackgroundBannerDirective.ts');
+require('components/common-layout-directives/common-elements/background-banner.directive.ts');
 require('pages/OppiaFooterDirective.ts');
