+++ core/templates/dev/head/pages/creator_dashboard/CreatorDashboard.ts
@@ -85,8 +85,8 @@ require('components/forms/HtmlSelectDirective.ts');
 require('services/AutoplayedVideosService.ts');
 // ^^^ this block of requires should be removed ^^^
 
-require('components/share/SharingLinksDirective.ts');
-require('components/background/BackgroundBannerDirective.ts');
+require('components/common-layout-directives/common-elements/sharing-links.directive.ts');
+require('components/common-layout-directives/common-elements/background-banner.directive.ts');
 require('components/summary_tile/CollectionSummaryTileDirective.ts');
 require('pages/exploration_player/PlayerConstants.ts');
 require('pages/exploration_editor/feedback_tab/ThreadTableDirective.ts');
