+++ core/templates/dev/head/pages/exploration_editor/ExplorationSaveAndPublishButtonsDirective.ts
@@ -16,7 +16,7 @@
  * @fileoverview Directive for the exploration save & publish buttons.
  */
 
-require('components/loading/LoadingDotsDirective.ts');
+require('components/common-layout-directives/common-elements/loading-dots.directive.ts');
 
 require('domain/utilities/UrlInterpolationService.ts');
 require('pages/exploration_editor/ChangeListService.ts');
