+++ core/templates/dev/head/pages/exploration_editor/ExplorationSaveService.ts
@@ -16,7 +16,7 @@
  * @fileoverview Service for exploration saving & publication functionality.
  */
 
-require('components/share/SharingLinksDirective.ts');
+require('components/common-layout-directives/common-elements/sharing-links.directive.ts');
 
 require('domain/exploration/StatesObjectFactory.ts');
 require('domain/utilities/UrlInterpolationService.ts');
