+++ core/templates/dev/head/pages/exploration_editor/translation_tab/TranslationTabDirective.ts
@@ -16,7 +16,7 @@
  * @fileoverview Directive for the translation tab.
  */
 
-require('components/attribution_guide/AttributionGuideDirective.ts');
+require('components/common-layout-directives/common-elements/attribution-guide.directive.ts');
 require(
   'pages/exploration_editor/translation_tab/StateTranslationDirective.ts');
 require(
