+++ core/templates/dev/head/pages/exploration_player/LearnerViewInfoDirective.ts
@@ -18,7 +18,7 @@
  */
 
 require('components/profile_link/ProfileLinkImageDirective.ts');
-require('components/share/SharingLinksDirective.ts');
+require('components/common-layout-directives/common-elements/sharing-links.directive.ts');
 require('filters/SummarizeNonnegativeNumberFilter.ts');
 require('filters/TruncateAndCapitalizeFilter.ts');
