+++ core/templates/dev/head/pages/learner_dashboard/LearnerDashboard.ts
@@ -16,8 +16,8 @@
  * @fileoverview Controllers for the creator dashboard.
  */
 
-require('components/background/BackgroundBannerDirective.ts');
-require('components/loading/LoadingDotsDirective.ts');
+require('components/common-layout-directives/common-elements/background-banner.directive.ts');
+require('components/common-layout-directives/common-elements/loading-dots.directive.ts');
 require('components/summary_tile/CollectionSummaryTileDirective.ts');
 require('components/summary_tile/ExplorationSummaryTileDirective.ts');
 require('filters/TruncateFilter.ts');
