+++ core/templates/dev/head/pages/practice_session/PracticeSession.ts
@@ -21,7 +21,7 @@ oppia.constant(
   'PRACTICE_SESSIONS_DATA_URL',
   '/practice_session/data/<topic_name>');
 
-require('components/background/BackgroundBannerDirective.ts');
+require('components/common-layout-directives/common-elements/background-banner.directive.ts');
 require('pages/question_player/QuestionPlayerDirective.ts');
 
 require('services/AlertsService.ts');
