+++ core/templates/dev/head/pages/thanks/Thanks.ts
@@ -16,7 +16,7 @@
  * @fileoverview Controllers for the 'thanks' page.
  */
 
-require('components/background/BackgroundBannerDirective.ts');
+require('components/common-layout-directives/common-elements/background-banner.directive.ts');
 require('pages/OppiaFooterDirective.ts');
 
 require('domain/utilities/UrlInterpolationService.ts');
