+++ core/templates/dev/head/pages/topic_editor/TopicEditorNavbarDirective.ts
@@ -16,7 +16,7 @@
  * @fileoverview Directive for the navbar of the topic editor.
  */
 
-require('components/loading/LoadingDotsDirective.ts');
+require('components/common-layout-directives/common-elements/loading-dots.directive.ts');
 
 require('domain/editor/undo_redo/UndoRedoService.ts');
 require('domain/topic/TopicRightsBackendApiService.ts');
