+++ core/templates/dev/head/pages/topics_and_skills_dashboard/TopicsAndSkillsDashboard.ts
@@ -16,7 +16,7 @@
  * @fileoverview Controllers for the topics and skills dashboard.
  */
 
-require('components/background/BackgroundBannerDirective.ts');
+require('components/common-layout-directives/common-elements/background-banner.directive.ts');
 require('pages/topics_and_skills_dashboard/SkillsListDirective.ts');
 require(
   'pages/topics_and_skills_dashboard/' +
